Trigger MELPA update
Just a dummy commit to trigger MELPA update one more time. I see that
the new recipe contains the `signs` folder, but for some reason the
folder was not packaged during the MELPA update.

So currently Pacmacs installed from MELPA crashes on death, because it
cannot find the `signs` folder. If the problem will persist even after
this dummy commit we will file an issue agains MELPA.