Start branch for #41 meta feature, retarget badges
Meta feature #41 requires lots of changes that are pretty risky. We
gonna develop it in a separate branch. All the subtasks related to #41
should be merged to this branch. Once the implementation is ready and
has sufficent performance, we gonna merge this branch to the
master. This makes it easier to decline everything related to the one
canvas technique if something goes wrong.